+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 ### Added
 - Added a second parameter to `Matter.component`, which allows specifying default component data.
--
+- Add `QueryResult:snapshot` to convert a `QueryResult` into an immutable list
+
 ### Changed
 - `queryChanged` behavior has changed slightly: If an entity's storage was changed multiple times since your system last observed it, the `old` field in the `ChangeRecord` will be the last value your system observed the entity as having for that component, rather than what it was most recently changed from.
 - World and Loop types are now exported (#9)